Switch to circleci debian stretch image
Using circleci's image allows us to be closer to normal open source configs and not worry about funny variances in our custom image. Updating to debian stretch removes the need for a bunch of packages, and also allows us to install osslsigncode from a package instead of building from source. Also, update to node 8 LTS because it's better.